example_with_logging/requirements.txt
cognite-logger>=0.5
